ref: constants and config moved to separate folder
