+++ app.js
@@ -3,16 +3,14 @@ const mongoose = require('mongoose');
 const helmet = require('helmet');
 const rateLimit = require('express-rate-limit');
 const cors = require('cors');
-const { MONGO_URL, PORT } = require('./config');
+const { MONGO_URL, PORT } = require('./config/config');
 const { errorsHandler } = require('./middlewares/errorsHandler');
+const { validationError } = require('./middlewares/validationError');
 const routes = require('./routes');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-const { errors, messages } = require('./constants');
+const { errors, messages, limiterConfig } = require('./config/constants');
 
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 100,
-});
+const limiter = rateLimit(limiterConfig);
 
 const app = express();
 app.use(helmet());
@@ -21,21 +19,16 @@ app.use(cors({
   methods: ['GET', 'POST', 'PATCH', 'PUT', 'DELETE'],
   allowedHeaders: ['Origin', 'Access-Control-Allow-Origin', 'X-Requested-With', 'Content-Type', 'Accept', 'Authorization'],
 }));
+
+app.use(requestLogger);
 app.use(limiter);
 app.use(express.json());
 
-app.use(requestLogger);
 app.use(routes);
 app.use(errorLogger);
 
 app.use(errorsHandler);
-app.use((err, req, res, next) => {
-  const { statusCode = 500, message } = err;
-  res.status(statusCode).send({
-    message: statusCode === 500 ? errors.serverError : message,
-  });
-  next();
-});
+app.use(validationError);
 
 async function main() {
   try {
