+++ config/constants.js
@@ -21,12 +21,18 @@ const errors = {
   authDataWasNotSend: 'Email or passwod was not send',
   authDataFailed: 'Email or password is not correct',
   authorizationRequired: 'Access denied, authorization required',
+  forbiddenError: 'Forbidden, 403',
   pageNotFound: 'Page not found, 404',
   serverError: 'Server error',
   connectionError: 'Conection failed',
   deleteMovieNotCorrectData: 'Inputed data is not correct',
 };
 
+const limiterConfig = {
+  windowMs: 15 * 60 * 1000,
+  max: 100,
+};
+
 module.exports = {
-  errors, messages, logs,
+  errors, messages, logs, limiterConfig,
 };
