+++ errors/ForbiddenError.js
@@ -1,0 +1,8 @@
+class ForbiddenError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 403;
+  }
+}
+
+module.exports = ForbiddenError;
