+++ errors/NotAuthorizedError.js
@@ -1,7 +1,7 @@
 class NotAuthorizedError extends Error {
   constructor(message) {
     super(message);
-    this.statusCode = 403;
+    this.statusCode = 401;
   }
 }
